Renames contactFormSection to ContactForm as component Uses correct label elements for contact form to surpress error.
